pyproject.toml
[tool.poetry]
name = "drf-homework"
version = "0.1.0"
description = ""
authors = ["Valera K <vk1337atliberty@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.5.3"
stripe = "^9.11.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
redis = "^5.0.6"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
